style: move logout button to the sidebar
