+++ imports/ui/components/AppSidebar.tsx
@@ -1,8 +1,13 @@
-import { LayoutDashboard, UserPen, Users } from "lucide-react";
+import { useQueryClient } from "@tanstack/react-query";
+import { LayoutDashboard, LogOut, UserPen, Users } from "lucide-react";
+import { Meteor } from "meteor/meteor";
+import { useNavigate } from "react-router-dom";
 
+import { Button } from "../elements/button";
 import {
   Sidebar,
   SidebarContent,
+  SidebarFooter,
   SidebarGroup,
   SidebarGroupContent,
   SidebarGroupLabel,
@@ -32,13 +37,22 @@ const items = [
 ];
 
 export function AppSidebar() {
+  const navigate = useNavigate();
+  const queryClient = useQueryClient();
+
+  const handleLogout = () => {
+    Meteor.logout();
+    queryClient.clear();
+    navigate(ROUTES.SIGN_IN);
+  };
+
   return (
     <Sidebar>
       <SidebarContent>
-        <SidebarGroup>
-          <SidebarGroupLabel>Application</SidebarGroupLabel>
-          <SidebarGroupContent>
-            <SidebarMenu className="font-semibold">
+        <SidebarGroup className="p-5">
+          <SidebarGroupLabel>Meteor SaaS Boilerplate</SidebarGroupLabel>
+          <SidebarGroupContent className="font-semibold mt-10">
+            <SidebarMenu>
               {items.map((item) => (
                 <SidebarMenuItem key={item.title}>
                   <SidebarMenuButton asChild>
@@ -53,6 +67,12 @@ export function AppSidebar() {
           </SidebarGroupContent>
         </SidebarGroup>
       </SidebarContent>
+      <SidebarFooter className="p-5">
+        <Button variant="outline" onClick={handleLogout} className="w-full shadow-sm">
+          <LogOut />
+          Logout
+        </Button>
+      </SidebarFooter>
     </Sidebar>
   );
 }
